requirements.txt
requests>=2.28.0
python-telegram-bot>=20.0
asyncio
